# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HACKDAY'S API",
        "description": "Documentation API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/candidats/store": {
            "post": {
                "tags": [
                    "Candidats"
                ],
                "summary": "Créer un nouveau candidat",
                "operationId": "storeCandidat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "email",
                                    "video"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "example": "0601020304"
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "example": "123 rue Exemple"
                                    },
                                    "ville": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "pays": {
                                        "type": "string",
                                        "example": "France"
                                    },
                                    "date_naissance": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-15"
                                    },
                                    "niveau_etude": {
                                        "type": "string",
                                        "example": "Bac+3"
                                    },
                                    "video": {
                                        "description": "Fichier vidéo du candidat",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidat enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidat enregistré avec succès."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Dupont"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "Jean"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jean.dupont@example.com"
                                                },
                                                "video": {
                                                    "type": "string",
                                                    "example": "videos/nom_fichier.mp4"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de validation incorrectes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/set-presence": {
            "post": {
                "tags": [
                    "Candidats"
                ],
                "summary": "Créer un nouveau candidat",
                "description": "Mettre à jour la présence d'un candidat en fonction de son email.",
                "operationId": "PresenceCandidat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean.dupont@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidat enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Candidat enregistré avec succès."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "nom": {
                                                    "type": "string",
                                                    "example": "Dupont"
                                                },
                                                "prenom": {
                                                    "type": "string",
                                                    "example": "Jean"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "jean.dupont@example.com"
                                                },
                                                "video": {
                                                    "type": "string",
                                                    "example": "videos/nom_fichier.mp4"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de validation incorrectes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The email field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}